Add the ability to check if a path exists
The new function apteryx_exists() checks if the specified path
exists.  If the path exists it returns true otherwise false.

This means that users of this function no longer need to call
apteryx_get() and free memory when all they don't care about the
value.  This removes the chance of a memory leak and allows the
caller to make a single line call instead of eight lines of code.