Better Lua detection.
This patch changes the testing for whether lua exists and whether it is available
as package "lua" or "lua5.2" covering both Ubuntu style and Gentoo style pkg-config
definitions. By changing to using the $(shell ... ) Makefile syntax, we gain that
the testing is only done once. This does however make the Makefile Gnu Make specific
for now.

Signed-off-by: Blair Steven <blair.steven@alliedtelesis.co.nz>