Alfred: Implement Index and fix the lua stack
Previously, the length of the stack was -1, so by removing set_global
on 'Apteryx', the length is now 0.
Index has been implemented to return a list of paths which exist
on that node.
Also have changed alfred_init and alfred_shutdown to act on the
global alfred_inst. This makes it easier for unit tests to be run.

Signed-off-by: Carl Smith <carl.smith@alliedtelesis.co.nz>