Add apteryx_get_tree_timeout function
This patch adds the function apteryx_get_tree_timeout which allows
the RPC timeout to be specified. This may be required if a tree takes
significantly longer to get than the default timeout of 1s.

The original apteryx_get_tree now wraps apteryx_get_tree_timeout and
passes in the default timeout.

A 'timeout' field has been added to the rpc_client struct, which is
assigned the desired value before connecting to the server.