Use separate timeout value for clients
When calling apteryx_get_tree, the RPC times out if getting the tree
takes longer than 1s. A solution is to have a separate define for
the client-side RPC timeout value.

This patch adds a separate define for the client RPC timeout value,
RPC_CLIENT_TIMEOUT_US. When apteryx_init is called, it sets the
timeout in the rpc_instance to this value, which is then used when
communicating with apteryxd.

The behaviour of apteryxd is unchanged; it still uses the original
timeout value.