Add the ability to check if a path exists and has a value
The new function apteryx_has_value() checks if the specified path
exists and has a value.  If the path exists and has a value it
returns true otherwise false.

This means that users of this function no longer need to call
apteryx_get() and free memory when all they don't care about the
value.  This removes the chance of a memory leak and allows the
caller to make a single line call instead of eight lines of code.