Exclude star nodes from search results
Currently if there is a field below a star node that is provided,
doing a search on the parent node will include the star node in the
search results.
e.g. If /a/b/*/f is provided, doing a search on /a/b/ will include
/a/b/* in the results.

This is undesirable because doing a search on an empty subtree should
return NULL.

This patch excludes paths containing star nodes when adding provided
paths to the search results.