Close the rpc client when the proxy is deconfigured
If the proxy is removed because the remote node went away,
then the node comes back, the first attempt to use the connection
would fail because we kept the connection.
So close the connection when the proxy is deconfigured,
if/when it is recreated the proxy connection will be re-established
on first use.

Signed-off-by: Carl Smith <carl.smith@alliedtelesis.co.nz>