Add option to create .run file once apteryxd is initialised
This patch adds a program option to apteryxd which allows the user to
specify a run file to create once apteryxd is initialised, e.g.
/var/run/apteryxd.run
This is useful for delaying the start-up of dependent processes such
as apteryx-sync.

A run file is not created unless the -r program option is used. Using
the -b option but omitting the -p option now does not create a default\
PID file.