rpc: when unproxy is called, only destroy the connection if it exists
Previously, if the connection didn't exist it would attempt to be
created which can take a long time (20 secs) to fail.
Since this code is only here to help cleanup when the target is
known to have gone away, only cleanup if we have an active connection
to the target.

Signed-off-by: Carl Smith <carl.smith@alliedtelesis.co.nz>