Check for previous unwatch
If unwatch is called twice for the same callback, the function/callback index
mapping will have been cleared. lua_tointeger will return 0, and apteryx_unwatch
will be called on callback 0. Regardless of the callback index the function was
registered with.

Checking for a non-nill value under the registry entry for the function fixes
this.