set: Don't prune on set to NULL
If set to NULL removes all children, then there is no way to
unset a value from a node without recreating all the children.
Make set to NULL only remove the value from the node and
leave the node around if it has children.
The timestamp on the node will be updated because it's value
has changed.