Use correct max ipv6 addr length in parse_url
The maximum ipv6 string represnetation length is 45 characters and
occurs with the form that represents the last 4 bytes as dotted decimal.

The length of the address field in the IPv6 TCP url parser was 48
characters which could overflow the host buffer which is only 46
characters long.

This has been changed to 45 to ensure this does not occur.