Gracefully handle fork
When a process linked to an initialized apteryx client library forks,
one of the threads and all of the internal data and socket handles are
copied to the new process.  All locks are left in their current state. This
means that in the child process, the state of the client is indeterminate.

Register pthread_atfork handlers to take the global lock and release it,
so that we at least have one lock we know is safe in the child process.
Also, set client data to NULL in the child process, so that it can no
longer interfere with any sockets, etc. created by the parent.  Currently,
if the child calls apteryx_shutdown, the parent process loses its callback
socket.

This breaks the test_double_fork unit test. That test expects to be able to
set/get after a fork without reinitialising.

Signed-off-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>