Use monotonic time for timestamps
Using the real clock time can cause problems for
Apteryx or any users of Apteryx that use
apteryx_timestamp when a change to the clock has
occurred. For example, syncer from apteryx-utils
can get stuck continously synching paths when the
clock shifts backwards or not sync at all if the
clock shifted forwards.

Signed-off-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>