Allow apteryx_query to return subtrees similar to apteryx_get_tree
If the root node in a query request is a longer path, rather than just
"/", use the full path as the root of the response tree, rather than
creating nodes for all of the path elements.

Previously it was possible to create a query like
 Root:/test/routing/ipv4/rib/1
 -> proto
 -> ifname
but the response would come back as
 Root:/test/routing/ipv4/rib/1 # in theory this should be "/"
 -> test
 --> routing
 ---> ipv4
 ----> rib
 -----> 1
 ------> proto = value
 ------> ifname = value

Now it will come back as:
 Root:/test/routing/ipv4/rib/1
 -> proto = value
 -> ifname = value

The existing behaviour still works the same if the query is passed in
with a "/" as the root.