Speed up apteryx_get_tree operations
A couple of changes to speed up apteryx_get_tree:

1) Hunt for refreshers directly, rather than searching for and running
indexers and providers before looking for refreshers. Only hunt for
refreshers / indexers / providers when we know there are some left down
this branch of the tree.

2) When traversing down the paths keep track of which types of callbacks
are present, and quit looking for callbacks that don't exist down a
branch. This is acheived with a new cb_exists function that returns true
if this path has a callback of the given type below it somewhere.