Delete correct callback when paths and fn are the same
If data has been provided then we need to compare that
as well. Currently used by LUA where the C fn is the
same, but the data (ref to the lua fn) are different.

Protect next_ref with the existing lock in case there
are two callbacks created at the same time.