apteryx_timestamp: support refreshers / providers
apteryx_timestamp returns a value that can be used to detect changes in
the database. When we have providers, a read of that value will always
be required, with refreshers this is required some period after the
previous update.

apteryx_timestamp will now return the current time when a provided value
timestamp is requested, and when an expired refreshed value is
requested. This value will be less than the subsequent data read, as is
required by the interface.