apteryx_refresh: Handle apteryx_refresh callbacks with a new function
apteryx_refresh_callback function type returns a uint64_t.
The call_callback function casts this to void *. On 32 bit systems, this
is problematic as half of the returned uint64_t is lost. Depending on
endianness, this may be the half with useful data.

This patch adds a new callback calling function for uint64_t return type.
If the callback is not found, it returns 0.