Ensure providers are called the minimum number of times
Ensure we only fetch information from the database / providers when
required. If a filter value is also a provided value we may need to call
it for each list entry at that level, but no more often than that.

Provided values that are to be filled in (specified with a null data
leaf node) will only be fetched when the filter for that list node
matches.