Make sure RPC is tidied up when shared library is unloaded
When a LUA script uses the Apteryx LUA module it is possible for
RPC messages to continue to arrive after the shared library has
been unloaded as LUA doesn't know how to shutdown the module.
This can lead to segment violations in LUA. A similar situation
is possible via apteryx-xml.

By flagging apteryx_shutdown_force as a destructor the RPC will
be correctly shutdown as the shared library is unloaded.