Add support for more control of set watch callbacks
If multiple YANG models configure the same data in a system then
a complex set of watches are required to keep the various processes
implementing the models synchronized. As an example of this consider the
implementation of the IETF and Openconfig interface models. If both
implementations watch the data of the other model (needed to keep the
configuration data synchronized), then if one model changes its data, the
other model will receive a watch callback and update its data to match.
This causes the first model to receive a watch callback about the change
in the second model. This easily leads to a storm of watch callbacks.

This change allows more options of how watch callbacks are generated
when data is set in database. Two new options are now possible when
setting data:-
1) Set data in the database but do not call any watch callbacks.
2) Set data in the database but block any callbacks to a specific watcher.