Lock the slow work thread context when modifying delayed work
When g_main_loop dispatches a timed callback it does so
with the glib main context unlocked. Which means that we
cannot safely restart the timer as we maybe inside the
slow work callback. Use the rpc lock to protect the g_main_loop
context in both the slow thread callbacks and when modifying
the timer in the rpc callback functions.