Faster library unload routine
When the library is unloaded lua requires that the threads are stopped -
this is done in two parts by stopping and joining any RPC client
threads, and stopping the thread pool.

The longer this process takes the more likely it is that one of the
worker threads will attempt to access routines from other libraries that
have unloaded and cleared their own memory. This can lead to undesired
log messages and / or processes exiting with other error codes.