Add check that a cb_info remains valid
If work is delayed (currently only done for watch_tree) and the watch
callback is removed while the work is in the queue, it can be used after
it has been freed. In this case the callback info structure has been
removed from the list - we can use this to determine its validity.