Make refresh more useful
Originally a refresher would be called for every request
if there was either an older child DB node or nothing
in the DB on that path at all.

Recently a change was made to only call a refresher once
in the timeout duration.

Now we record the last refreshed path so that we can refresh
multiple specific paths that match the refresher path.