Fix refreshing paths for wildcards
Do not call refresher if we have unresolved wildcards.
Pass an actual wildcard node when recursing _refresh_paths
so that strcmp matches properly.
Fixup test asssertions for cases that should not have
called a refresher and cases that were missing the final /.
Add a bunch of test cases for root path queries and refreshers.