Fix local refresher for statistics
We should record the timestamp and last path on any
local refreshers.

We dont need to specify the CAS timestamp when adding stats to the DB,
but using the wrong timestamp fn is even worse.

Currently the only local refresher is for statistics and this fix
avoids the stats refresher being thrashed.