Return DB timestamp for refreshed paths
Previously we alway returned "now" if there was a chance
of a path being refreshed. That was because refreshed paths
always had their timestamps updated to "now" even if the
refresher decided not to update anything so that we did not
call the refresher again for stale child nodes. Now that we
do not use the DB timestamps to decide whether to call the
refreshers anymore we can fix this odd behaviour.

The change in behaviour is;
1. A call to a refresher does not result in all children of
the path having their timestamps set to "now" even if they
were not modified.
2. A call to get a timestamp for a path will call refreshers
to make sure the timestamp is up to date.