Check apteryx_set_tree with multiple values
Ensure a client setting up multiple values for the same key in a tree
ends up with the last value actually being used. There are ways of doing
this setup which will end up with the wrong value being set (requiring a
mix of methods to create the tree), but this will work when
APTERYX_VALUE / APTERYX_NODE is used the same way multiple times to
create a tree with duplicate nodes.