Update apteryx query assertions
When showing an incomplete nat rule router would crash, due to an
invalid memory access. This was caused by having a validity check on an
entity that does not exist resulting in an apteryx node with null data.
This has been remedied by updating the assertion in the apteryx query func
to return NULL when the node data is NULL. Preventing an invalid memory
access when destroying the node.