Store all strings in a cache
Apteryxd can end up storing a lot of duplicated strings. This commit
adds a cache to remove this duplication - it is called the same way
strdup / free would be and will take care of allocating, reference
counting and releasing strings.

This cache has no locking, so needs to be called under an external lock.
At the moment this used solely in code inside the db_lock in database.c